pyproject.toml
[project]
name = "bd-web"
version = "0.1.0"
description = ""
authors = [{ name = "Camilo Paez", email = "camilopaez.dev@gmail.com" }]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv (>=0.9.9,<0.10.0)",
    "pyodbc (>=5.2.0,<6.0.0)",
    "flask (>=3.1.0,<4.0.0)",
    "mysql-connector-python (>=9.2.0,<10.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
